btrfs: Factor out the main deletion process from btrfs_ioctl_snap_destroy()
Factor out the secand half of btrfs_ioctl_snap_destroy() as
btrfs_delete_subvolume(), which performes some subvolume specific checks
before deletion (1. send is not in progress. 2. the subvolume
is not the default subvolume. 3. the subvolume does not contain other
subvolumes) and actual deletion process. btrfs_delete_subvolume() requires
inode_lock for both @dir and inode of @dentry.

Note that call of d_delete() is not included in btrfs_delete_subvolume()
as this function will also be used by btrfs_rmdir() to delete an empty
subvolume and in that case d_delete() is called in vfs layer.

As a result, btrfs_unlink_subvol() and may_destroy_subvol()
become static functions. No functional change happens.